requirements.txt
termcolor==1.1.0
python-apt==1.8.4.3 